Allow mio_new_mio() to have size 0
Consider for instance the following HTML code where a subparser is called
on the contents of the <script> tags.

<script language="Javascript" src="./test.js"></script>
<script language="JavaScript">
...
</script>

Since the first <script> is empty, mio_new_mio() is called with 0 size
but with the current semantics of the API it creates MIO until the end of
the file so the whole source file to EOF is parsed. While it seems
no extra tags are generated in this case, the parsing happens
unnecessarily.

Instead of using 0, make size signed and use -1 when we want to create a
MIO till the end of the source MIO.

+++ main/mio.h
@@ -123,7 +123,7 @@ MIO *mio_new_memory (unsigned char *data,
 					 MIOReallocFunc realloc_func,
 					 MIODestroyNotify free_func);
 
-MIO *mio_new_mio    (MIO *base, long start, size_t size);
+MIO *mio_new_mio    (MIO *base, long start, long size);
 MIO *mio_ref        (MIO *mio);
 
 int mio_free (MIO *mio);
